requirements.txt
flask~=3.0.2
langchain==0.1.6
langchain-community==0.0.19
better_profanity~=0.7.0
pgvector~=0.2.5
psycopg2-binary~=2.9.9
tiktoken~=0.6.0
python-dotenv~=1.0.1
langchain-openai~=0.0.6
langchainhub~=0.1.14
pytz~=2024.1
requests~=2.31.0